feat(v3): Remove the test workflow file and update the submodule and go.mod file
Deleted the test.yml file in the.github/workflows directory, which was used to automatically trigger test jobs. In addition, the.gitmodules file has been updated to add paths, urls, and branch information for the v2 and v3 submodules. This indicates that the code base may be preparing for a new release or reorganizing the code structure. At the same time, the module path in the go.mod file changes from a version-numbered path to a non-version-numbered path, which may mean that the developer intends to reset the codebase to the main version. A lot of code has been removed from the entslog package, including the SlogDriver and SlogTx structures and related functions. These deletions imply that the log-driven and transaction logging implementations may change significantly or be replaced. Also removed are the handle.go and option.go files, which contain implementation details for logging and configuration options. These changes may mean significant changes to the library's internal architecture and apis to provide a cleaner or more flexible logging mechanism.

Overall, this code change reflects a major refactoring of the codebase to improve logging and code organization, possibly in preparation for the upcoming release.